requirements.txt
aioconsole==0.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-cors==0.8.1
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.4.0
attrs==25.3.0
av==14.4.0
bidict==0.23.1
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
dnspython==2.7.0
frozenlist==1.7.0
google-crc32c==1.7.1
h11==0.16.0
idna==3.10
ifaddr==0.2.0
multidict==6.6.3
numpy==2.2.6
opencv-python==4.12.0.88
propcache==0.3.2
pycparser==2.22
pyee==13.0.0
pylibsrtp==0.12.0
pyOpenSSL==25.1.0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
requests==2.32.4
simple-websocket==1.1.0
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.1
